pyproject.toml
[project]
name = "price_tracker_backend"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "arainfaisal826", email = "arainfaisal826@gmail.com" }
]
dependencies = [
    "fastapi>=0.111.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "SQLAlchemy>=2.0.0",  # Explicitly adding SQLAlchemy
    "asyncpg>=0.29.0",    # For PostgreSQL async driver
    "sqlalchemy[asyncio]>=1.4.30", # For SQLAlchemy asyncio support
    "celery>=5.3.0",
    "redis>=4.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/app"]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "httpx>=0.23.0",
    "pytest-asyncio>=0.20.0", # Required for async pytest functions
]
